feat: 대회탭 알람 기능 추가
